changed traditional functions to arrow functions
