+++ .env.example
@@ -1,4 +1,4 @@
-REACT_APP_HOST = http://localhost
+REACT_APP_HOST = http://localhost:8000
 REACT_APP_PORT = 8000
 REACT_APP_BIO_PATH = /api/bio/
 REACT_APP_BLOG_PATH = /api/blog/
