+++ src/services/LogoutRequest.js
@@ -1,29 +1,12 @@
-import Cookies from 'js-cookie';
+import '../utils/options'
 
-const getOptions = () => {
-
-    const options = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Authorization': null,
-        'X-CSRFToken': Cookies.get('csrftoken')
-      },
-      credentials: 'include',
-      body: "",
-    };
-
-    return options
-  }
-
-function LogoutRequest() {
+const LogoutRequest = () => {
 
     let options = getOptions()
 
         const env = process.env;
 
-        const url = `${env.REACT_APP_HOST}:${env.REACT_APP_PORT}${env.REACT_APP_LOGOUT_PATH}`
+        const url = `${env.REACT_APP_HOST}${env.REACT_APP_LOGOUT_PATH}`
 
         return fetch(url, options)
         .then(response => response.json())
