+++ src/services/RegisterRequest.js
@@ -1,31 +1,13 @@
-import Cookies from 'js-cookie';
+import '../utils/options'
 
-
-const getOptions = credentials => {
-
-    const options = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        'Authorization': null,
-        'X-CSRFToken': Cookies.get('csrftoken')
-      },
-      body: JSON.stringify(credentials),
-    };
-
-    return options
-  }
-
-
-function registerRequest(credentials) {
+const registerRequest = (credentials) => {
    
 
     let options = getOptions(credentials)
 
         const env = process.env;
 
-        const url = `${env.REACT_APP_HOST}:${env.REACT_APP_PORT}${env.REACT_APP_REGISTER_PATH}`
+        const url = `${env.REACT_APP_HOST}${env.REACT_APP_REGISTER_PATH}`
 
         return fetch(url, options)
         .then(response => response.json())
